+++ lib/presentation/router/app_router.dart
@@ -1,10 +1,11 @@
+import 'package:fintek/presentation/screens/home_screen.dart';
 import 'package:flutter/material.dart';
 
 import '../../core/exceptions/route_exception.dart';
 import '../screens/splash_screen.dart';
 
 class AppRouter {
-  static const String splash = '/';
+  static const String splash = '/', home = 'home';
 
   const AppRouter._();
 
@@ -13,6 +14,10 @@ class AppRouter {
       case splash:
         return MaterialPageRoute(
           builder: (_) => const SplashScreen(),
+        );
+      case home:
+        return MaterialPageRoute(
+          builder: (_) => const HomeScreen(),
         );
       default:
         throw const RouteException('Route not found!');